feat: feign 라이브러리 관련 spring boot 버전 하향(3.3.5) - 테스트 로직 수정 - fastAPI fallback 방식 변경(retry 유지 -> songService에서 fallback 처리)
