+++ src/main/java/com/movewave/emotion/service/EmotionServiceImpl.java
@@ -20,13 +20,6 @@ import lombok.RequiredArgsConstructor;
 @Service
 @RequiredArgsConstructor
 public class EmotionServiceImpl implements EmotionService {
-    /** 기본 감정 상태 */
-    private static final String NEUTRAL_EMOTION = "중립";
-    /** 기본 신뢰도 값 */
-    private static final double DEFAULT_CONFIDENCE = 0.0;
-    /** 기본 키워드 목록 */
-    private static final List<String> DEFAULT_KEYWORDS = List.of("편안한 음악");
-    
     /** Flask API 클라이언트 */
     private final FlaskApiClient flaskApiClient;
     
@@ -45,12 +38,7 @@ public class EmotionServiceImpl implements EmotionService {
         backoff = @Backoff(delay = 1000)   // 재시도 간 간격(ms)
     )
     public EmotionResponse analyzeEmotion(String text, String type) {
-        try {
-            return flaskApiClient.analyzeEmotion(createRequestBody(text, type));
-        } catch (Exception e) {
-            log.warn("Flask API 실패 → fallback 사용: {}", e.getMessage());
-            return fallbackEmotion();
-        }
+        return flaskApiClient.analyzeEmotion(createRequestBody(text, type));
     }
     
     /**
@@ -62,9 +50,5 @@ public class EmotionServiceImpl implements EmotionService {
      */
     private Map<String, String> createRequestBody(String text, String type) {
         return Map.of("text", text, "type", type);
-    }
-
-    private EmotionResponse fallbackEmotion() {
-        return new EmotionResponse(NEUTRAL_EMOTION, DEFAULT_CONFIDENCE, DEFAULT_KEYWORDS);
     }
 }
