+++ src/test/java/com/movewave/emotion/service/EmotionServiceImplTest.java
@@ -35,7 +35,6 @@ class EmotionServiceImplTest {
     private static final String TEST_EMOTION = "기쁨";
     private static final double TEST_CONFIDENCE = 0.85;
     private static final List<String> TEST_KEYWORDS = List.of("신나는", "댄스");
-    private static final String EMOTION_API_PATH = "/api/emotion/predict";
 
     @Test
     @DisplayName("정상적인 감정 분석 응답 처리")
@@ -43,37 +42,14 @@ class EmotionServiceImplTest {
         // given
         EmotionResponse expected = new EmotionResponse(TEST_EMOTION, TEST_CONFIDENCE, TEST_KEYWORDS);
 
-        when(flaskApiClient.post(
-                eq(EMOTION_API_PATH),
-                any(Map.class),
-                eq(EmotionResponse.class)
-        )).thenReturn(CompletableFuture.completedFuture(expected));
+        when(flaskApiClient.analyzeEmotion(any(Map.class))).thenReturn(expected);
 
         // when
-        CompletableFuture<EmotionResponse> result = emotionService.analyzeEmotion(TEST_TEXT, TEST_TYPE);
+        EmotionResponse response = emotionService.analyzeEmotion(TEST_TEXT, TEST_TYPE);
 
         // then
-        assertThat(result).isCompleted();
-        EmotionResponse response = result.join();
         assertThat(response.prediction()).isEqualTo(TEST_EMOTION);
         assertThat(response.confidence()).isEqualTo(TEST_CONFIDENCE);
         assertThat(response.keywords()).isEqualTo(TEST_KEYWORDS);
-    }
-
-    @Test
-    @DisplayName("Flask 서버 통신 실패 시 폴백 메서드 호출")
-    void fallbackEmotion_Success() {
-        // given
-        RuntimeException exception = new RuntimeException("Flask 서버 통신 실패");
-
-        // when
-        CompletableFuture<EmotionResponse> result = emotionService.fallbackEmotion(TEST_TEXT, TEST_TYPE, exception);
-
-        // then
-        assertThat(result).isCompleted();
-        EmotionResponse response = result.join();
-        assertThat(response.prediction()).isEqualTo("중립");
-        assertThat(response.confidence()).isEqualTo(0.0);
-        assertThat(response.keywords()).containsExactly("편안한 음악");
     }
 }
